feat(src): apply clang-tidy fix for bugprone-implicit-widening-of-multiplication-result check

+++ src/member.c
@@ -6,6 +6,7 @@
 #include "utils.h"
 #include "validate.h"
 #include <ctype.h>
+#include <stddef.h>
 #include <stdio.h>
 #include <string.h>
 #include <time.h>
@@ -1693,7 +1694,7 @@ void memberViewKicked(AppDatabase *db) {
 
     for (int line = 0; line < lines; line++) {
       char buf[25] = {0};
-      strncpy(buf, reason + (line * chunk), (size_t)chunk);
+      strncpy(buf, reason + (ptrdiff_t)(line * chunk), (size_t)chunk);
 
       printf(COLOR_CYAN "  " LINE_V COLOR_RESET);
       if (line == 0) {
